+++ src/algorithm/dp/coinChange/coinChange.ts
@@ -3,25 +3,28 @@ export function coinChange(coins: number[], amount: number): number {
     return -1;
   }
 
+  if (amount === 0) {
+    return 0;
+  }
+
   const dp = Array(amount + 1)
     .fill(-1);
-  dp[0] = 0;
 
   for (let i = 0; i < coins.length; i++) {
     if (coins[i] === amount) {
       return 1;
     }
 
-    if (coins[i] < amount) {
-      dp[coins[i]] = 1;
-    }
+    dp[coins[i]] = 1;
   }
 
   for (let i = 1; i <= amount; i++) {
     for (let j = 0; j < coins.length; j++) {
-      if (i >= coins[j] && dp[i - coins[j]] !== -1) {
-        if (dp[i] === -1 || dp[i] > dp[i - coins[j]] + 1) {
+      if (i >= coins[j] && dp[i - coins[j]] > 0) {
+        if (dp[i] === -1) {
           dp[i] = dp[i - coins[j]] + 1;
+        } else {
+          dp[i] = Math.min(dp[i], dp[i - coins[j]] + 1);
         }
       }
     }